Python: When converting to JSON, use OrderedDict to preserve field order.
Only some of our structs need to preserve field order, so this is
broader than strictly necessary.

In the long run, we probably also want to switch to something that
returns a string generator instead of a single big string.  I think this
will make a difference for API calls with large JSON results, like
"list_folder".

Reviewed By: kelkabany
Differential Revision: https://tails.corp.dropbox.com/D84159