Route I/O can reference nullable types. A variant can be a nullable type.
Summary:
This introduces "null" into our JSON serialized format.

In the case of the nullable variant, we wouldn't be able to substitute it
with an absent JSON field since we need the tag, so I'm okay with it.

At some point, @kannan and @guido can continue to discuss whether lists should
be able to contain nullable types, but let's make that separate from this.

Test Plan: Already had tests for this.

Reviewers: kannan, guido

Reviewed By: guido

Subscribers: kannan, guido

Differential Revision: https://tails.corp.dropbox.com/D86634