A struct field default can now reference a union symbol variant.
Summary:
This gives you the ability to specify UploadMode.add (union variant)
as the default for a struct field of type UploadMode. This was one of
the features you requested in your review of the Python v2 SDK.

In the interest of the mini hackathon this week, I'm just going to land this.

Test Plan: Added tests

Reviewers: guido

Subscribers: guido

Differential Revision: https://tails.corp.dropbox.com/D90209