Bad defaults now print an error message rather than dumping a Python traceback.
Summary:
When a default didn't pass validation for a type, babel used to return an
ugly Python traceback. Now it shows a descriptive error message similar to
all the other error messages it displays to the user (with line numbers and
hints).

For the mini-hackathon, I'm going to land this now. Feel free to comment on the
diff later.

Reviewers: guido

Subscribers: guido

Differential Revision: https://tails.corp.dropbox.com/D90213