Moved Python generator out of the babelapi Python package.
Summary:
It's still in the repository, but it's expected that the babelapi
user will reference it by its full path rather than just "python".

I want to bring this functionality back, but when I do, I'll be doing
it in a slightly different way.

Test Plan: Re-ran tests and used babelapi.

Reviewers: kannan, guido

Reviewed By: guido

Subscribers: jlee, rpearl, guido

Differential Revision: https://tails.corp.dropbox.com/D92244