Changes necessary so routes.babelg.py can generate imports for foreign data types.
Summary:
This is needed so I can make D108646 work (which has a route whose
error type is imported from a different spec).
I will send you the change to routes.babelg.py in the server repo
that needs this separately.

Test Plan: Generated code in the server repo for D108646.

Reviewers: kelkabany

Reviewed By: kelkabany

Subscribers: guido

Differential Revision: https://tails.corp.dropbox.com/D110455